requirements.txt
moviepy==1.0.3
requests==2.31.0
pytube==15.0.0
selenium==4.15.2
python-decouple==3.8
aiogram==3.2.0
boto3==1.33.5
botocore==1.33.5